requirements.txt
# Torch
#--find-links https://download.pytorch.org/whl/torch_stable.html
#torch==1.4.0+cu100
#torchvision==0.11.2+cu100
#torchaudio==0.10.1+cu100
#setuptools==59.5.0

Pillow
ninja
tqdm
opencv-python
scikit-image
numpy

tensorboard

# for face alignment
tensorflow
#keras
#bz2
dlib
scipy

matplotlib
pprintpp
